+++ src/z/zinitn.c
@@ -17,6 +17,6 @@ hebi_zinitn(size_t count, struct hebi_integer r[count])
 		r[i].hz_resv = 0;
 		r[i].hz_used = 0;
 		r[i].hz_sign = 0;
-		r[i].hz_allocid = (int)(intptr_t)id;
+		r[i].hz_allocid = id;
 	}
 }
